requirements.txt

Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
django-filter==23.5
django-cors-headers==4.3.1
qrcode==7.4.2
djangorestframework-simplejwt